Update test_PSO_implementation.ipynb
using Manopt: i think this will import all the methods of the class

called particle_swarm! directly

Since "!" mark is used the values in the "pts" changed
First set of values in the pts variable should be the optimum
get_solver_result(res) gives the optimum.

Please run the score function on the optimum point and check.